+++ test/Fixtures/Mariadb/FullDatabases/CreateOrReplace/create_or_replace.sql
@@ -1,0 +1,15 @@
+CREATE OR REPLACE TABLE `account` (
+  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
+  `email` varchar(255) NOT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `uq_email` (`email`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+
+CREATE OR REPLACE TABLE `account_role` (
+  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
+  `account_id` int(10) unsigned NOT NULL,
+  `role` varchar(32) NOT NULL,
+  PRIMARY KEY (`id`),
+  KEY `fk_account_id` (`account_id`),
+  CONSTRAINT `fk_account_id` FOREIGN KEY (`account_id`) REFERENCES `account` (`id`) ON DELETE CASCADE
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;